refac: move results to separate package
