+++ src/main/java/medicaldoctor/LoginController.java
@@ -5,7 +5,7 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
-import medicaldoctor.backend.LoginResult;
+import medicaldoctor.backend.results.LoginResult;
 import medicaldoctor.backend.LoginService;
 
 public class LoginController {
