+++ src/main/java/medicaldoctor/backend/LoginService.java
@@ -1,5 +1,6 @@
 package medicaldoctor.backend;
 
+import medicaldoctor.backend.results.LoginResult;
 import medicaldoctor.core.AppSession;
 import medicaldoctor.core.DatabaseScope;
 import medicaldoctor.entities.User;
